requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-multipart==0.0.6
pdfplumber==0.10.3
pandas==2.1.3
plotly==5.18.0
streamlit==1.29.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
apscheduler==3.10.4
plaid==17.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9 